requirements.txt
# Python dependencies

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Drivers
asyncpg==0.29.0
neo4j==5.15.0

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI/ML libs
openai==1.3.7
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2

# Vector support
pgvector==0.2.4

# HTTP
httpx==0.25.2
aiohttp==3.9.1

# Data processing
pandas==2.1.4
python-dateutil==2.8.2

# Utils
python-dotenv==1.0.0
loguru==0.7.2
structlog==23.2.0

# Dev dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Testing
httpx==0.25.2
faker==20.1.0
